Default `write_row` to use `write_string`.
Instead of `write`. 

This allows us to keeping leading zeros on values like `0001`. Using `write` converts this to a number, like `1`, instead of keeping the leading zeros.

If this were to be merged back into the main repo this could be set as an option like `strings: true` or by setting a global config to always use `write_string` when calling `write_row`. But we can do that later if there's interest. For now, we'll use this fork.