Disable custom formatting of `Time` objects.
Complaint from Tim at SoCalGas. The old way was I guess better. I sent him a video of how to look at the custom formatting and how to change it to either date or time or both.

But honestly, Excel's native support for time zones is laughable so it might be best just to treat Time as a String instead, like we were doing before.

_I also removed the commented-out 24-hour support. Won't work because what date would you choose. Blah.