Don't convert to Dates to Times.
We're seeing a very odd thing happen when in development the dates show up properly but in production (using the same version/commit of the gem) it's showing the time and timezone as a String. No idea why that would be the case.

### Development
`2018-07-05`

### Production
`2018-07-05T00:00:00+00:00`

Of course, because this is only happening in Production, this is actually pretty tough to test without re-deploying a change and see if it makes it better.

My thinking here is that not converting it to a `Time` before passing to `write_date_time` might prevent the time and timezone from being rendered.

I'll give it a shot. If this doesn't work, I'm just going to render Dates as a String and be done with it.