Don't rely on `write_date_time` and instead calculate the days since Excel epoch ourselves and just use write_number.
Again, trying to fix an issue that only occurs in production where the date is showing with time and timezone.

My thinking here is that there is too much magic going on with `write_date_time`. And we need to pass in a time for that method to work so instead of doing that, let's do things ourselves and just pass the days since Excel epoch in as a number with a Date format.

In development, this does the correct thing. The raw field is just an integer and it is formatted as a Date correctly, allowing you to change the format from "2019-01-01" to "January 1st, 2019" very easily.

Going to deploy this to Prod and see if it fixes it there too.