Support formatting of 24 hour times.
Has to be in the exact format of "hh:mm:ss". Nothing else will work.

It uses the following REGEX to determine if it's in a proper 24 hour format and it is a valid time:

```
(?:[01]\d|2[0123]):(?:[012345]\d):(?:[012345]\d)
```

If it's a propery 24 hour time then it calculates the fraction of the day that it represents. For example, 18:00:00 is 3/4 or 0.75 of a day so the number it provides to Excel is 0.75 and it uses a format of "hh:mm:ss" to format it in.

But, because Excel has the raw fraction as the cell value, it can format it in any kind of time format that you want, including "h:mm AM/PM". Very clever stuff.